move sanitise and introduce exteded browser auth interface
