+++ auth/auth_test.go
@@ -114,3 +114,33 @@ func TestSave(t *testing.T) {
 		})
 	}
 }
+
+func Test_sanitize(t *testing.T) {
+	type args struct {
+		workspace string
+	}
+	tests := []struct {
+		name    string
+		args    args
+		want    string
+		wantErr bool
+	}{
+		{"not a URL", args{"blahblah"}, "blahblah", false},
+		{"url slash", args{"https://blahblah.slack.com/"}, "blahblah", false},
+		{"url no slash", args{"https://blahblah.slack.com"}, "blahblah", false},
+		{"url no schema slash", args{"blahblah.slack.com/"}, "blahblah", false},
+		{"url no schema no slash", args{"blahblah.slack.com"}, "blahblah", false},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := sanitize(tt.args.workspace)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("sanitize() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			if got != tt.want {
+				t.Errorf("sanitize() got = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
